check mo_pubs's existence before migrate (#20184)
check mo_pubs's existence before migrate

Approved by: @qingxinhome, @zhangxu19830126, @sukki37, @daviszhen, @LeftHandCold
