+++ pkg/bootstrap/versions/utils.go
@@ -1,0 +1,78 @@
+// Copyright 2021 Matrix Origin
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//	http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+package versions
+
+import (
+	"time"
+
+	"github.com/matrixorigin/matrixone/pkg/common/pubsub"
+	"github.com/matrixorigin/matrixone/pkg/container/types"
+	"github.com/matrixorigin/matrixone/pkg/container/vector"
+	"github.com/matrixorigin/matrixone/pkg/util/executor"
+)
+
+const (
+	getPubInfosSql = "select pub_name, database_name, database_id, table_list, account_list, created_time, update_time, comment from mo_catalog.mo_pubs"
+)
+
+var GetPubInfos = func(txn executor.TxnExecutor, accountId uint32, accountName string) (pubInfos []*pubsub.PubInfo, err error) {
+	var exist bool
+	if exist, err = CheckTableDefinition(txn, accountId, "mo_catalog", "mo_pubs"); err != nil || !exist {
+		return
+	}
+
+	// select from old mo_pubs table, which located in each account
+	res, err := txn.Exec(getPubInfosSql, executor.StatementOption{}.WithAccountID(accountId))
+	if err != nil {
+		return
+	}
+	defer res.Close()
+
+	res.ReadRows(func(rows int, cols []*vector.Vector) bool {
+		for i := 0; i < rows; i++ {
+			var pubInfo pubsub.PubInfo
+			pubInfo.PubAccountName = accountName
+			pubInfo.PubName = cols[0].GetStringAt(i)
+			pubInfo.DbName = cols[1].GetStringAt(i)
+			pubInfo.DbId = vector.GetFixedAtWithTypeCheck[uint64](cols[2], i)
+			pubInfo.TablesStr = cols[3].GetStringAt(i)
+			pubInfo.SubAccountsStr = cols[4].GetStringAt(i)
+			pubInfo.CreateTime = vector.GetFixedAtWithTypeCheck[types.Timestamp](cols[5], i).String2(time.Local, cols[5].GetType().Scale)
+			if !cols[6].IsNull(uint64(i)) {
+				pubInfo.UpdateTime = vector.GetFixedAtWithTypeCheck[types.Timestamp](cols[6], i).String2(time.Local, cols[6].GetType().Scale)
+			}
+			pubInfo.Comment = cols[7].GetStringAt(i)
+			pubInfos = append(pubInfos, &pubInfo)
+		}
+		return true
+	})
+	return
+}
+
+// GetAllPubInfos returns map[pubAccountName#pubName] -> pubInfo
+var GetAllPubInfos = func(txn executor.TxnExecutor, accNameInfoMap map[string]*pubsub.AccountInfo) (map[string]*pubsub.PubInfo, error) {
+	allPubInfos := make(map[string]*pubsub.PubInfo)
+	for _, accountInfo := range accNameInfoMap {
+		pubInfos, err := GetPubInfos(txn, uint32(accountInfo.Id), accountInfo.Name)
+		if err != nil {
+			return nil, err
+		}
+
+		for _, pubInfo := range pubInfos {
+			allPubInfos[accountInfo.Name+"#"+pubInfo.PubName] = pubInfo
+		}
+	}
+	return allPubInfos, nil
+}
