+++ pkg/common/pubsub/utils.go
@@ -26,10 +26,6 @@ import (
 	"github.com/matrixorigin/matrixone/pkg/util/executor"
 )
 
-const (
-	getPubInfosSql = "select pub_name, database_name, database_id, table_list, account_list, created_time, update_time, comment from mo_catalog.mo_pubs"
-)
-
 func InSubMetaTables(m *plan.SubscriptionMeta, tableName string) bool {
 	if strings.ToLower(m.Tables) == TableAll {
 		return true
@@ -136,49 +132,4 @@ var GetAccounts = func(txn executor.TxnExecutor) (nameInfoMap map[string]*Accoun
 		return true
 	})
 	return
-}
-
-var GetPubInfos = func(txn executor.TxnExecutor, accountId uint32, accountName string) (pubInfos []*PubInfo, err error) {
-	// select from old mo_pubs table, which located in each account
-	res, err := txn.Exec(getPubInfosSql, executor.StatementOption{}.WithAccountID(accountId))
-	if err != nil {
-		return
-	}
-	defer res.Close()
-
-	res.ReadRows(func(rows int, cols []*vector.Vector) bool {
-		for i := 0; i < rows; i++ {
-			var pubInfo PubInfo
-			pubInfo.PubAccountName = accountName
-			pubInfo.PubName = cols[0].GetStringAt(i)
-			pubInfo.DbName = cols[1].GetStringAt(i)
-			pubInfo.DbId = vector.GetFixedAtWithTypeCheck[uint64](cols[2], i)
-			pubInfo.TablesStr = cols[3].GetStringAt(i)
-			pubInfo.SubAccountsStr = cols[4].GetStringAt(i)
-			pubInfo.CreateTime = vector.GetFixedAtWithTypeCheck[types.Timestamp](cols[5], i).String2(time.Local, cols[5].GetType().Scale)
-			if !cols[6].IsNull(uint64(i)) {
-				pubInfo.UpdateTime = vector.GetFixedAtWithTypeCheck[types.Timestamp](cols[6], i).String2(time.Local, cols[6].GetType().Scale)
-			}
-			pubInfo.Comment = cols[7].GetStringAt(i)
-			pubInfos = append(pubInfos, &pubInfo)
-		}
-		return true
-	})
-	return
-}
-
-// GetAllPubInfos returns map[pubAccountName#pubName] -> pubInfo
-var GetAllPubInfos = func(txn executor.TxnExecutor, accNameInfoMap map[string]*AccountInfo) (map[string]*PubInfo, error) {
-	allPubInfos := make(map[string]*PubInfo)
-	for _, accountInfo := range accNameInfoMap {
-		pubInfos, err := GetPubInfos(txn, uint32(accountInfo.Id), accountInfo.Name)
-		if err != nil {
-			return nil, err
-		}
-
-		for _, pubInfo := range pubInfos {
-			allPubInfos[accountInfo.Name+"#"+pubInfo.PubName] = pubInfo
-		}
-	}
-	return allPubInfos, nil
 }
